+++ ext/stub/oo.zep.c
@@ -199,7 +199,7 @@ PHP_METHOD(Stub_Oo, testInstance9)
 	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
 
-	ZEPHIR_CALL_CE_STATIC(&o, stub_oo_oodynamica_ce, "getnew", NULL, 0);
+	ZEPHIR_CALL_CE_STATIC(&o, stub_oo_oodynamica_ce, "getNew", NULL, 0);
 	zephir_check_call_status();
 	RETURN_CCTOR(&o);
 }
@@ -214,7 +214,7 @@ PHP_METHOD(Stub_Oo, testInstance10)
 	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
 
-	ZEPHIR_CALL_CE_STATIC(&o, stub_oo_oodynamicb_ce, "getnew", NULL, 0);
+	ZEPHIR_CALL_CE_STATIC(&o, stub_oo_oodynamicb_ce, "getNew", NULL, 0);
 	zephir_check_call_status();
 	RETURN_CCTOR(&o);
 }